data description work (commit 2)
Revised some of my prior notes and completed the data description by using the players.csv file. Will be deleting the read_csv() and summarize() code used to gather the data description, as they rely on an external file, and the project proposal does not allow use of those (must run independently).