questions and start of exploratory data work (commit 3)
added the broad question, specific questions, and explanations. added back code used to read in dataset and summarize for single statistical variables in part 1 as I misunderstood the project guidelines and thought I was not supposed to load in the dataset at any point in order for the file to run without additional files.